requirements.txt
google_api_python_client==2.39.0
oauth2client==4.1.3
python_dateutil==2.8.2
PyYAML==6.0
